+++ generator-maven-plugin/generator-plugin/src/test/java/uk/gov/justice/maven/generator/io/files/parser/generator/GenerateGoalProcessorTest.java
@@ -27,8 +27,6 @@ import java.nio.file.Path;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;