+++ job-manager/job-executor/src/test/java/uk/gov/moj/cpp/task/execution/JobExecutorTest.java
@@ -15,9 +15,8 @@ import static org.mockito.Mockito.when;
 import static uk.gov.moj.cpp.jobstore.api.task.ExecutionInfo.executionInfo;
 import static uk.gov.moj.cpp.jobstore.api.task.ExecutionStatus.COMPLETED;
 import static uk.gov.moj.cpp.jobstore.api.task.ExecutionStatus.INPROGRESS;
+import static uk.gov.moj.cpp.jobstore.persistence.Priority.HIGH;
 
-import java.util.List;
-import java.util.Optional;
 import uk.gov.justice.services.common.util.UtcClock;
 import uk.gov.moj.cpp.jobstore.api.task.ExecutionInfo;
 import uk.gov.moj.cpp.jobstore.persistence.Job;
@@ -26,6 +25,8 @@ import uk.gov.moj.cpp.task.extension.SampleTask;
 import uk.gov.moj.cpp.task.extension.TaskRegistry;
 
 import java.time.ZonedDateTime;
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import javax.json.JsonObject;
@@ -252,7 +253,7 @@ public class JobExecutorTest {
         final UUID jobId = randomUUID();
         final JsonObject jobData = mock(JsonObject.class);
         final NotSupportedException notSupportedException = new NotSupportedException();
-        final Job job = new Job(jobId, jobData, "taskName", new UtcClock().now(), empty(), empty(), 0);
+        final Job job = new Job(jobId, jobData, "taskName", new UtcClock().now(), empty(), empty(), 0, HIGH);
         final JobExecutor jobExecutor = createJobExecutor(job);
 
         when(taskRegistry.getTask(eq("taskName"))).thenReturn(empty());
@@ -270,7 +271,7 @@ public class JobExecutorTest {
         final JsonObject jobData = mock(JsonObject.class);
         final NotSupportedException notSupportedException = new NotSupportedException();
         final SystemException systemException = new SystemException();
-        final Job job = new Job(jobId, jobData, "taskName", new UtcClock().now(), empty(), empty(), 0);
+        final Job job = new Job(jobId, jobData, "taskName", new UtcClock().now(), empty(), empty(), 0, HIGH);
         final JobExecutor jobExecutor = createJobExecutor(job);
 
         when(taskRegistry.getTask(eq("taskName"))).thenReturn(empty());
@@ -297,6 +298,7 @@ public class JobExecutorTest {
                 .withNextTaskStartTime(nextTaskStartTime)
                 .withNextTask("taskName")
                 .withRetryAttemptsRemaining(retryAttemptsRemaining)
+                .withPriority(HIGH)
                 .build();
     }
 }