+++ job-manager/jobstore-persistence/src/main/java/uk/gov/moj/cpp/jobstore/persistence/Priority.java
@@ -1,0 +1,7 @@
+package uk.gov.moj.cpp.jobstore.persistence;
+
+public enum Priority {
+    HIGH,
+    MEDIUM,
+    LOW;
+}